third_party/blink/web_tests/svg/dynamic-updates/SVGRectElement-dom-height-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'height' attribute of the SVGRectElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rectElement.getAttribute('height') is "50"
PASS rectElement.getAttribute('height') is "200"
PASS successfullyParsed is true

TEST COMPLETE
